requirements.txt
python-docx==0.8.11
requests==2.31.0
keyring==24.2.0